refactor: Update validation constraints for song upload DTOs
This commit updates the validation constraints for the `UploadSongDto` and `ThumbnailData` DTOs. The `title` property in `UploadSongDto` now has a maximum length of `UploadConst.title.maxLength`, and the `originalAuthor` property has a maximum length of `UploadConst.originalAuthor.maxLength`. Similarly, the `description` property in `UploadSongDto` has a maximum length of `UploadConst.description.maxLength`. In the `ThumbnailData` DTO, the `zoomLevel` property now has a maximum value of `ThumbnailConst.zoomLevel.max` and a minimum value of `ThumbnailConst.zoomLevel.min`. Additionally, the `startTick` property has a default value of `ThumbnailConst.startTick.default`, and the `startLayer` property has a default value of `ThumbnailConst.startLayer.default`. Lastly, the `backgroundColor` property in `ThumbnailData` now has a default value of `ThumbnailConst.backgroundColor.default`.

Co-authored-by: Bentroen <29354120+Bentroen@users.noreply.github.com>