feat: Add SOUND_LIST constant to InstrumentPicker component
This commit adds the SOUND_LIST constant to the InstrumentPicker component in order to utilize the latest version of the sound list. By importing the SOUND_LIST constant from the shared features/sounds module, the component can access the updated sound list and ensure that it is up to date. This change improves the functionality of the InstrumentPicker component.