fix: pad custom ins array with empty strings on upload, backend and edit
Having `undefined` values on the array would mess up the form schema validation, which only allows strings. The sounds array was being padded neither on the frontend, nor in the backend (should've been in both), causing less instruments to be stored than there were instruments in the song. Once the song was loaded in the edit form, not enough array positions would be filled to warrant every custom instrument being initialized with an empty string, meaning if they were not touched by the user, the form would be in an invalid/error state. To prevent this all, we make sure to:

- initialize the 'custom instruments' form with empty strings the length of the instruments in the song;
- pad the array when saving the song in the backend to have the length of the instruments in the song;
- initialize the 'custom instruments' dropdown values to have the length of the loaded song's instruments.