Refactor: Refactor User and Song Controllers and Services
This commit refactors the user and song controllers and services to use the UserService.verifyUser(user) static method to verify the userDocument obtained from the @GetRequestToken() decorator.

This change reduces the amount of redundant code in the controllers and services that was previously used solely for user verification and consolidates the logic within the UserService.verifyUser(user) method.