fix: upload complete modal not showing, song details overflow
The line of code to open the modal was inadvertently commented out in commit 2606d21f1ad49a237d3ea5d97c2d4e9710c454fa, which was unrelated to that but committed in a bit of a hurry (probably transitioning to a different machine) while experimenting with unrelated changes. Oops.